Only map keys when g:dwm_map_keys is 1
It is a good idea to give the user the possibility to override the default key
mappings. With this commit the user can set g:dwm_map_keys to 0 and call the dwm
functions himself. If the variable is not set at all it is assumed to be 1.
The solution is taken from a blog post by Steve Losh's [1].

[1] http://stevelosh.com/blog/2011/09/writing-vim-plugins/#mapping-keys-the-right-way